uhd: remove integer conversions and carry floats throughout
The UHD driver will perform an internal float conversion
from the over-the-wire type, which is then output through
the API as integer, which is later converted in the radio
interface back to float. This patch removes some of the
type conversion madness at the expense of doubling I/O
bandwidth at the lower transceiver layer, but this
appears to to be worthwhile tradeoff when squeezing
out performance.

Obviously, this breaks a number of interfaces and is
incompatible with non-UHD builds.

Signed-off-by: Thomas Tsou <ttsou@vt.edu>